+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+- Adds [`Range.compareBeginToEnd`][]
+
+[`Range.compareBeginToEnd`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.compareBeginToEnd
+
 ## 1.5.0 (Jul 12, 2016)
 - Adds [`Range.prototype.valueOf`][] to get a more primitive representation of a range.  
   Useful with [Egal.js][egal] or other libraries that compare value objects by their `valueOf` output.